adoption dan commisions

+++ database/migrations/2025_10_20_150721_create_commisions_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->string('category', 50);
             $table->text('description');
             $table->date('deadline')->nullable();
-            $table->decimal('price', 10, 2);
+            $table->float('price'); 
             $table->enum('payment_status', ['pending', 'dp', 'paid', 'refunded'])->default('pending');
             $table->enum('progress_status', ['pending', 'accepted', 'in_progress_sketch', 'in_progress_coloring', 'review', 'revision', 'completed', 'cancelled'])->default('pending');
             $table->timestamps();
